Add files via upload
Deciding a place for one’s new start up or establishing a new unit of an existent business can be
a difficult task. A significant number of factors have to be considered in making such an
important decision. This project gives an ability, for decision makers, to arrive at an informed
decision on which industry yields the best in which state, in which county and in which city.

County Business Patterns (CBP), ZIP Code Business Patterns (ZBP) are an annual series that
provide sub-national economic data by industry. These programs cover most of the U.S.
economy and feature industry and geographic statistics which supplement those published in
the Economic Census. Data are published at the U.S. level and by State, County, Metropolitan
area, ZIP code, and Congressional District. All data are classified by an industry code (NAICS)
and can be viewed with employment-size class breakouts by establishment, and by legal form
of organization at some geographic levels.